+++ assets/appFramework/angular/generate/guard/unit/mochachai/{GEN_NAME_SNAKE}.spec.ts
@@ -1,8 +1,7 @@
 /**
  * Tests for {GEN_NAME_PASCAL} guard.
  */
-import * as chaiModule from "chai";
-const chai = (chaiModule as any).default || chaiModule;
+import {SYNTHETIC_IMPORT}chai from "chai";
 import { {GEN_NAME_PASCAL}Guard } from "{GEN_UNIT_TEST_RELATIVE}{GEN_NAME_SNAKE}{ADDITIONAL_EXTENSION}";
 
 describe("{GEN_NAME_PASCAL}Guard", () => {