+++ assets/appFramework/shared/test/unit/src/sourceLanguage/typescript/mochachai/app.spec.ts
@@ -1,8 +1,7 @@
 /**
  * Tests for App.
  */
-import * as chaiModule from "chai";
-const chai = (chaiModule as any).default || chaiModule;
+import {SYNTHETIC_IMPORT}chai from "chai";
 import { App } from "../../../src/app";
 
 describe("App", () => {