+++ test/unit/src/pipelineSteps/moduleType/systemJs.spec.ts
@@ -64,6 +64,24 @@ describe("SystemJs", () => {
         });
     });
 
+    describe("initialise", () => {
+        it("can be called with not matching condition", async () => {
+            engineVariablesStub.syntheticImport = "foo";
+            const obj = new SystemJs();
+            const res = await obj.initialise(loggerStub, fileSystemMock, uniteConfigurationStub, engineVariablesStub, false);
+            Chai.expect(res).to.be.equal(0);
+            Chai.expect(engineVariablesStub.syntheticImport).to.be.equal("foo");
+        });
+
+        it("can be called with matching condition", async () => {
+            engineVariablesStub.syntheticImport = "foo";
+            const obj = new SystemJs();
+            const res = await obj.initialise(loggerStub, fileSystemMock, uniteConfigurationStub, engineVariablesStub, true);
+            Chai.expect(res).to.be.equal(0);
+            Chai.expect(engineVariablesStub.syntheticImport).to.be.equal("");
+        });
+    });
+
     describe("configure", () => {
         it("can be called with no configurations", async () => {
             const obj = new SystemJs();